leetcode/hot100/hash/group_anagrams/sort_slice/map_sort_slice.go
package sort_slice

import "sort"

/*
给你一个字符串数组，请你将 字母异位词 组合在一起。可以按任意顺序返回结果列表。
字母异位词 是由重新排列源单词的所有字母得到的一个新单词。

输入: strs = ["eat", "tea", "tan", "ate", "nat", "bat"]
输出: [["bat"],["nat","tan"],["ate","eat","tea"]]
*/

// GroupAnagrams 把字符串里的字母排序，作为键，把字母异位词的slice作为值
func GroupAnagrams(strs []string) [][]string {
	mp := make(map[string][]string)
	for _, str := range strs {
		s := []byte(str)
		// sort.Slice会把对第一个参数的排序结果存回第一个参数
		// 所以既然后面还需要使用str，需要用str的复制，这里转为字符切片
		sort.Slice(s, func(i, j int) bool {
			return s[i] < s[j]
		})
		sortedStr := string(s)
		mp[sortedStr] = append(mp[sortedStr], str)
	}
	ans := make([][]string, 0, len(mp))
	for _, v := range mp {
		ans = append(ans, v)
	}
	return ans
}
